# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Autonomous Shopping AI Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: brand
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user's wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{id}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        price: { type: number }
        originalPrice: { type: ["number", "null"] }
        imageUrl: { type: ["string", "null"] }
        description: { type: string }
        specs: { type: object, additionalProperties: true }
        rating: { type: number }
        reviewCount: { type: integer }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        tags: { type: array, items: { type: string } }
        aiScore: { type: ["number", "null"] }
        createdAt: { type: string }
      required: [id, name, brand, category, price, description, rating, reviewCount, inStock, isFeatured, tags]
    ProductInput:
      type: object
      properties:
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        price: { type: number }
        originalPrice: { type: ["number", "null"] }
        imageUrl: { type: ["string", "null"] }
        description: { type: string }
        specs: { type: object, additionalProperties: true }
        rating: { type: number }
        reviewCount: { type: integer }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        tags: { type: array, items: { type: string } }
      required: [name, brand, category, price, description, rating, reviewCount, inStock, isFeatured, tags]
...
    WishlistItem:
      type: object
      properties:
        id: { type: integer }
        sessionId: { type: string }
        productId: { type: integer }
        product:
          $ref: "#/components/schemas/Product"
        addedAt: { type: string }
      required: [id, sessionId, productId, product]
    WishlistItemInput:
      type: object
      properties:
        productId: { type: integer }
        sessionId: { type: string }
      required: [productId, sessionId]
...