# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputSpecs } from './productInputSpecs';

export interface ProductInput {
  name: string;
  brand: string;
  category: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  description: string;
  specs?: ProductInputSpecs;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isFeatured: boolean;
  tags: string[];
}