# lib/api-zod/src/generated/types/wishlistItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface WishlistItemInput {
  productId: number;
  sessionId: string;
}